Redoing file structure
