--- src/ui/components/chunks/mod.rs
@@ -1,3 +1,0 @@
-pub mod chatting;
-pub mod help;
-pub mod states;
